fix: 🐛 lazy load images, lazy load carousel items
BREAKING CHANGE: 🧨 None

✅ Closes: [WIN-835]
