+++ package.json
@@ -115,6 +115,7 @@
     "react-hook-form": "^7.34.2",
     "react-image-gallery": "1.2.8",
     "react-image-lightbox": "^5.1.4",
+    "react-lazy-load-image-component": "^1.5.6",
     "react-leaflet": "^4.0.0",
     "react-router-dom": "^6.3.0",
     "react-scripts": "5.0.1",
